requirements.txt
Django==2.2.10
django-filer
celery
redis
Landsat578